+++ src/com/zs/gitpractice/model/EmployeeType.java
@@ -1,4 +1,4 @@
-package com.zs.gitpractice.p1;
+package com.zs.gitpractice.model;
 
 public enum EmployeeType {
     FullTime, ParTime, FixedTerm, Probably, Outworker, DailyHired, WeeklyHired;
